상속 리뷰

+++ Java20250306_함수/src/ch01/ex01/함수매개변수있고반환값있는경우03.java
@@ -15,14 +15,14 @@ package ch01.ex01;
 public class 함수매개변수있고반환값있는경우03 {
 
 	public static void main(String[] args) {
-        int result =  sum(10, 20);  //이 코드 실행화면 결과는
+        double result =  sum(10, 20);  //이 코드 실행화면 결과는
         // int result = 30;        --> 이렇게 나온다.
 		System.out.println(result);          
 		System.out.println("프로그램 종료!!");
 	}
 	
 	//함수 정의
-	public static int sum(int n1, int n2) {
+	public static double sum(double n1, double n2) {
 		//코드 구현
 		System.out.println("void sum(int n1, int n2) 함수 호출!!");
